Renaming src/objectify.js
